6-size.c

+++ 0x00-hello_world/6-size.c
@@ -9,7 +9,7 @@ int main(void)
 	int b;
 	long int c;
 	long long int d;
-	long float f;
+	float f;
 
 printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
 printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
